Separate cosmos module in base and java parts
