+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos-java/src/main/java/com/microsoft/azure/toolkit/intellij/cosmos/code/function/CosmosDBFunctionPathAnnotator.java
