+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/resources/META-INF/azure-intellij-plugin-cosmos.xml
@@ -4,16 +4,10 @@
         <registryKey key="azure.toolkit.cosmos_cassandra.dbtools.enabled"
                      description="Enable DB Tools related features of Azure cosmos DB API for Cassandra"
                      restartRequired="true" defaultValue="false"/>
-        <completion.contributor id="azCosmosDBCompletion" language="JAVA" order="after azStorageStringLiteral" implementationClass="com.microsoft.azure.toolkit.intellij.cosmos.code.function.AzureCosmosDBFunctionAnnotationCompletionContributor"/>
-        <psi.referenceContributor language="JAVA" implementation="com.microsoft.azure.toolkit.intellij.cosmos.code.function.AzureCosmosDBResourceReferenceContributor"/>
-        <annotator language="JAVA" implementationClass="com.microsoft.azure.toolkit.intellij.cosmos.code.function.CosmosDBFunctionPathAnnotator"/>
     </extensions>
     <extensions defaultExtensionNs="com.microsoft.tooling.msservices.intellij.azure">
         <explorerNodeProvider implementation="com.microsoft.azure.toolkit.ide.cosmos.CosmosNodeProvider"/>
         <actions implementation="com.microsoft.azure.toolkit.ide.cosmos.CosmosActionsContributor"/>
         <actions implementation="com.microsoft.azure.toolkit.intellij.cosmos.IntelliJCosmosActionsContributor"/>
-        <connectorResourceType implementation="com.microsoft.azure.toolkit.intellij.cosmos.connection.CassandraCosmosDBAccountResourceDefinition"/>
-        <connectorResourceType implementation="com.microsoft.azure.toolkit.intellij.cosmos.connection.MongoCosmosDBAccountResourceDefinition"/>
-        <connectorResourceType implementation="com.microsoft.azure.toolkit.intellij.cosmos.connection.SqlCosmosDBAccountResourceDefinition"/>
     </extensions>
 </idea-plugin>
