+++ PluginsAndFeatures/azure-toolkit-for-intellij/settings.gradle.kts
@@ -24,6 +24,7 @@ include("azure-intellij-plugin-appservice-java")
 include("azure-intellij-plugin-arm")
 include("azure-intellij-plugin-applicationinsights")
 include("azure-intellij-plugin-cosmos")
+include("azure-intellij-plugin-cosmos-java")
 include("azure-intellij-plugin-cognitiveservices")
 include("azure-intellij-plugin-monitor")
 include("azure-intellij-plugin-containerregistry")
